etc: Some fine-tuning parameters
After some playing with the parameters I found, that the temperature
profile cannot be believed for politropic planets.
The reason is simple: for stable sphere one needs higher K, wich promo-
es high temperatures. The entirity of the planet is not politropic! But
this is still good for starting guesses at least.

TODO:
- Opacity wrapper class designing
- Refactor: Outsourcing instance-dependent details for Build_envelope
into Planet